Include support link and consistent layout
This top navigation is shared by 3 separate apps, we want it to feel
like you are browsing one app.

All 3 applications should have the same navigation items available for
easy browsing and should look consistent.

This sets a max width on desktop to match the layout of the admin and
product page.