restructure admin requirements to split out recommendations
There are certain steps that aren't actually required to get set up on GovWifi.

To avoid the misconception of needing to potentially do lots of work, make it clear
what is required to use GovWifi, and what should be followed to provide good QoS.

use of Static IP addresses has also been reworded, to explain the static IP address
is only required for the RADIUS servers.

Note that the requirement to allow VPN services has been left in as a requirement,
as their use is ubiquitous around most departments, and so should continue to function