Refactor way of deploying app
Currently we're dealing with potential secrets leakage management... The
way we're deploying things is printing out the credentials to the
pipeline. This isn't the worst thing ever, as Concourse is only
accessible by people within the Bardeen network.

This however, is still potential threat... It could lead to GovWifi
product page being overwritten with no audit of when, why or who...

Refactoring the tasks will fix that issue, but bring other problems
along the way...

This way, it is incredibly slow to actually perform the same task. There
is however room for improvement, as the main priority is the security
incident.