feat: install a cookie banner on the site
This unfortunately requires a substantial amount of changes since the
shared GOV.UK Tech Docs gem is only a Middleman template and can't be
overriden easily.

So introduce a package.json including the GOVUK Frontend + the shared
GovWifi frontend package, and backport some of the scrips of GovWifi
product page, namely copy-assets.rb which makes the assets available
for Middleman to pickup later.

Then just call `GovWifi.cookies.checkCookiePolicy()` and the job is
done.