Move backup plan content to 'Set up GovWifi'
When reading through the docs, I think this is actually a better place for it. It's more like a thing you have to do when setting GovWifi up initially, rather than something you have to do on an ongoing basis once it's already been set up. (which is the 'Manage' page). It also follows on nicely from the 'incident alerts' section.