Move 'tech docs' explanation to the top
Moved the tech docs bit to the top due to the duplication between this page and the 'Offer' page. I think the duplication is fine for now because of the different audiences, but this is a 'just in case' someone does read both and gets a bit confused about how the pages differ.